feat: recover from panic from call to NewAttribute()
This change adds a recovery mechanism to the pkcs11.NewAttribute()
function to prevent from bringing down the application. Instead, the
error is captured and returned in an error variable. This required
all of the callers to now handle the returned error.

In addition, this change cleaned up the way the templates were
being duplicated to create the public/private sets. It should be
obvious why we are cloning the AttributeSets.