Add function for accessing underlying PKCS#11 keys
Add a function to allow users of this package to access underlying PKCS#11-isms so that custom extensions can be built.
For example the new function could be used to wrap/export a key or unwrap/import a key using the handle associated
with an existing private or secret key.